+++ src/pipdot/__main__.py
@@ -2,6 +2,7 @@
 Generate a GraphViz dot file representing installed PyPI distributions
 """
 
+import json
 import site
 import sys
 from argparse import ArgumentParser, FileType
@@ -9,6 +10,7 @@ from functools import lru_cache, partial
 from itertools import chain
 from os import path
 from pathlib import Path
+from subprocess import check_output
 
 try:
     from argparse import BooleanOptionalAction  # type: ignore
@@ -35,8 +37,15 @@ def get_args():
         )
     )
     parser.add_argument(
+        '--python', '-P', type=str,
+        help='Report the distributions for this python executable. '
+             '"--path/-p" option will not affect if this option was set. '
+             'The program check specified python\'s "sys.path" by an actual execution, so you can not use it in a container.'
+    )
+    parser.add_argument(
         '--path', '-p', action='append',
-        help='If path is set, only report the distributions present at the specified location. '
+        help='Report the distributions present at the specified location. '
+             'This option will not affect if "--python/-P" was set. '
              'This option can be specified multiple times for more than one locations.'
     )
     parser.add_argument(
@@ -145,13 +154,21 @@ def perform(args):
     from packaging.requirements import Requirement  # type: ignore
     from packaging.utils import canonicalize_name  # type: ignore
 
+    if args.python:
+        output = check_output([
+            args.python, '-c', 'import json, sys; json.dump(sys.path, sys.stdout)'
+        ])
+        sys_paths = json.loads(output)
+        args.path = [p for p in sys_paths if path.isdir(p)]
+
     kdargs = dict()
     if args.path:
         kdargs.update(path=args.path)
     dists = list(importlib_metadata.distributions(**kdargs))
 
     requires_extras = lru_cache(maxsize=None)(
-        partial(_get_requires_extras, dists))
+        partial(_get_requires_extras, dists)
+    )
     installed = lru_cache(maxsize=None)(partial(_installed, dists))
 
     context = {